Add ViewExt.kt
