+++ sample/src/main/java/jp/co/cyberagent/android/gpuimage/sample/utils/ViewExt.kt
@@ -1,0 +1,31 @@
+package jp.co.cyberagent.android.gpuimage.sample.utils
+
+import android.view.View
+import androidx.core.view.ViewCompat
+
+inline fun View.doOnLayout(crossinline action: (view: View) -> Unit) {
+    if (ViewCompat.isLaidOut(this) && !isLayoutRequested) {
+        action(this)
+    } else {
+        doOnNextLayout { action(it) }
+    }
+}
+
+inline fun View.doOnNextLayout(crossinline action: (view: View) -> Unit) {
+    addOnLayoutChangeListener(object : View.OnLayoutChangeListener {
+        override fun onLayoutChange(
+            view: View,
+            left: Int,
+            top: Int,
+            right: Int,
+            bottom: Int,
+            oldLeft: Int,
+            oldTop: Int,
+            oldRight: Int,
+            oldBottom: Int
+        ) {
+            view.removeOnLayoutChangeListener(this)
+            action(view)
+        }
+    })
+}
